Created Upload Page
I created an upload page that only takes fit files under 10 Mb of size. It then takes the user on a viewer page where it (temporarily) writes the dates of all the waypoints in the file uploaded. If the file produce an error, the user is taken back to the upload page accompagnied by an error.