Added interactible test graph with altitude, heart rate, and speed
I used chart.js to create a graph with 3 y axes, one for altitude, one for heart rate and the last one for the speed. This is only a test but I already applied a size to the graph and customized the little over toolbox interface. Also, I only show one point out of ten (one every ~ten seconds) to improve performance.